refactor(tile_stitch): rename tile_cache module to dispatch
Module holds Morton ordering + the dispatch budget, no cache. Rename
tile_cache.py -> dispatch.py and its two tests; update the two imports.
